check if it  contains the given value or not

+++ Arrays/AverageValue.java
@@ -5,11 +5,11 @@ public class AverageValue {
     public static void main(String[] args) {
         int [] array = new int[] {1,2,3,4,5};
         int sum = 0;
-        int average = 0;
+        double average = 0;
         for(int i = 0; i < array.length; i++) {
             sum += array[i];
         }
-        average = sum / array.length;
+        average = (double) sum / array.length;
 
         System.out.println("Given Array is >> " + Arrays.toString(array) );
         System.err.println("The Average is >>> " + average);
